Change sentDate from verzendDatum to datumBeschikbaar
The parameter verzendDatum that we get from Kalliope is a date that
doesn't specify the hours/minutes/... (ex: "2019-06-17"). But we need
the exact time the message is sent to be able to have the right order
for the messages in a conversation. datumBeschikbaar specify the time
and is related to when the message became available for the API, it's
the parameter we'll use.