Remove the graph constraint in a SELECT query
There is no need to spec the graph here, as we are interested in
knowing wether the inzending has been sent, not where it has been
updated.

In current context of loket, we have the vendor-data-distributon
graph, that might not always contain up to date information about the
sync state with kalliope.